Part 2 of Python 2 to 3 Conversion
This part deals with the difference in sockets between Python 2 and 3.
In python 2, it is perfectly fine to send a string via a socket like so.
However, sockets in Python 3 can only send byte strings.  To handle this,
the only change I made was to encode a string before it was sent, and decode
the string just after it is received.